refactor: change modal service name file and add it on public api ts
