+++ projects/ion/src/public-api.ts
@@ -26,3 +26,4 @@ export * from './lib/card/card.component';
 export * from './lib/simple-menu/simple-menu.component';
 export * from './lib/modal/component/modal.component';
 export * from './lib/info-badge/info-badge.component';
+export * from './lib/modal/modal.service';
