+++ stories/Modal.stories.mdx
@@ -7,7 +7,7 @@ import dedent from 'ts-dedent';
 import { ButtonComponent } from '../projects/ion/src/lib/button/button.component.ts';
 import { IonIconComponent } from '../projects/ion/src/lib/icon/icon.component.ts';
 import { IonModalComponent } from '../projects/ion/src/lib/modal/component/modal.component';
-import { IonModalService } from '../projects/ion/src/lib/modal/ion-modal.service';
+import { IonModalService } from '../projects/ion/src/lib/modal/modal.service';
 import { OpenModalButtonComponent } from '../projects/ion/src/lib/modal/mock/open-modal.mock.component';
 import { SelectMockComponent } from '../projects/ion/src/lib/modal/mock/select-mock.component';
 import { InputMockComponent } from '../projects/ion/src/lib/modal/mock/input.mock.component.ts';
